radar_annotation_system/requirements.txt
# 核心依赖
numpy>=1.21.0
pandas>=1.3.0
opencv-python>=4.5.0
Pillow>=8.3.0

# 3D可视化
open3d>=0.13.0
plotly>=5.0.0
matplotlib>=3.4.0

# 深度学习框架
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0

# 目标检测和分割
segment-anything
grounding-dino-py
transformers>=4.20.0

# UI框架
streamlit>=1.20.0
gradio>=3.0.0

# 数据处理
scipy>=1.7.0
scikit-learn>=1.0.0
h5py>=3.1.0

# 配置和工具
pyyaml>=6.0
tqdm>=4.60.0
click>=8.0.0

# 开发工具
pytest>=6.0.0
black>=21.0.0
flake8>=3.9.0